requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
requests==2.32.4
python-dotenv==1.1.1
